requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.5
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
slowapi==0.1.9
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.24.3
scikit-learn==1.3.2
aiofiles==23.2.1
httpx==0.25.2
pydantic[email]
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1
xlwt==1.3.0
redis==5.0.1
apscheduler==3.10.4
celery==5.3.4
openpyxl==3.1.2
xlsxwriter==3.1.9
